requirements.txt
requests
pyedhrec
beautifulsoup4
numpy
sentence-transformers
tqdm
torch
transformers
networkx
flask
scikit-learn
umap-learn
matplotlib